+++ src/DI/Definition/Source/AnnotationBasedAutowiring.php
@@ -7,7 +7,6 @@ use DI\Annotation\Injectable;
 use DI\Definition\AutowireDefinition;
 use DI\Definition\EntryReference;
 use DI\Definition\Exception\AnnotationException;
-use DI\Definition\Exception\DefinitionException;
 use DI\Definition\ObjectDefinition;
 use DI\Definition\ObjectDefinition\MethodInjection;
 use DI\Definition\ObjectDefinition\PropertyInjection;
@@ -261,7 +260,7 @@ class AnnotationBasedAutowiring implements DefinitionSource, Autowiring
             $annotation = $this->getAnnotationReader()
                 ->getClassAnnotation($class, 'DI\Annotation\Injectable');
         } catch (UnexpectedValueException $e) {
-            throw new DefinitionException(sprintf(
+            throw new AnnotationException(sprintf(
                 'Error while reading @Injectable on %s: %s',
                 $class->getName(),
                 $e->getMessage()