+++ tests/IntegrationTest/Definitions/DecoratorDefinitionTest.php
@@ -3,7 +3,7 @@
 namespace DI\Test\IntegrationTest\Definitions;
 
 use DI\ContainerBuilder;
-use Interop\Container\ContainerInterface;
+use Psr\Container\ContainerInterface;
 
 /**
  * Test decorator definitions.
@@ -104,7 +104,7 @@ class DecoratorDefinitionTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException \DI\Definition\Exception\DefinitionException
+     * @expectedException \DI\Definition\Exception\InvalidDefinition
      * @expectedExceptionMessage Entry "foo" decorates nothing: no previous definition with the same name was found
      */
     public function test_decorate_must_have_previous_definition()